refactor: create wrapper for airport list
