+++ src/Components/AirportsList/types.ts
@@ -1,0 +1,6 @@
+import { AirportsListTypes } from '@redux/api/type';
+
+export type AirportsListProps = {
+  pending: boolean;
+  airportsListData: AirportsListTypes[];
+};
